Finished the tower objects until modifiers.
Arrow projectile is also now smaller and fits better.
Removed collision from the arrow projectile.
Added sound effects to the towers when they shoot, the volume is controlled via the TowerSoundEffectVolume property in the Help class.
A sound effect for the arrow tower and the alchemy tower have been added.
Added four towers that shoot at the first object found with the tag "Debug" to demonstrate the sound + projectiles. These can be removed.