Fixed bugs, added visualizer toggle, and made some stuff more effecient.
Made the circle class more effecient by only updating when needed.
Made the grid visualizer more effecient by reusing existing grid tiles.
Made the grid visualizer piece opaque.
Added the ability to make projectiles track their target.
Added a toggle for the grid visualizer in the toolbar.
Moved the ray casting from the grid class to the help class.
Fixed minor bug in the PlayerManager class when assigning the best path for a given grid.
Added a DEBUG button alongisde the tower and creature buttons to force an object to walk along the best path of a grid.
Added commented out code in the Astar algorithm that makes it more effecient, but currently has bugs.