Created a simple basic "AI".
Created a ValidPosition method in Help.cs to determine if a given x and y coordinate on a grid can have a tower placed on it.
Moved the tower placement code from TowerSpawn over to PlayerManager so it can be used for other players (IE: AI).
Deactivated the point light from the archer tower prefab.
All players now receive their income when the player's timer elapses.
Income timer is now 20 seconds.